Added assetsdir and moved convertDate function to timeSlotUtil
